Add functional communication between Python and Java
Adds the SquareNum.py file.

In VirtualBoardController, creates a SquareNum object from SquareNum.py.
This object takes in a numeric value and returns the square of that
number.

This exercise is a proof of concept for passing values into a Python
file from Java code and recieving usable data from that Python code and
using that data in more Java.