Adds a little gpio reader and fixes buttons issues
The first steps of adding the gpio reader functionality.

Fixes an issue with the new HBoxes, which weren't being added to the
vboxes array as they were not VBoxes.  So instances of VBoxes are
replaced with Pane classes, so both H and VBoxes can be used in the same
hash.  The buttons work now.