Updating Button/LED test to use a VirtualRelay
I hadn't realized that conditions were VR's so I've made a quick test to
see how reading those rather than from the ConditionMap works, and I
think it works really well.  Next step is to implement this for all
VirtualRelays and see if I can tie it together with a simpler
implementation of Button updating.