Updating VirtualSwitches
Updates VirtualSwitch objects to be more closely related to their Pane
counterparts, storing buttons and a SwitchState value.

Adds a new object, VirtualSwitchHandler, to be used to better store and
interact with VirtualSwitches.

These changes are the continuation in a series of attempts to improve
abstraction in the project, and increase overall efficiency.