First part of putting together a process to simply change a button and have the whole system update.
That's a long commit message but it's what's happening.  We're going to
have the system entirely work just for the Lock Rough Valve.  So that
and the values related to it are the only thing kept in this branch for
right now.

I have the GPIOReader working.  Next step is getting the relays to
update based on those virtual values.